Attempt fixing argon2 hashing
It seems like something changed with workerd that caused
the `await wasi.start()` method to start throwing errors like
"The script will never generate a response"

Maybe wasi.start is no longer a promise due to new wasm
optimizations? Just a guess. Having it await the wasi start
along with the stderr and stdout streams seems to resolve the error.